feat: add updateMetadataCallback mecanism in MetadataProvider
